refactor: CacheService dependencyies
