+++ dristi/lib/src/core/styles/texts.dart
@@ -1,4 +1,4 @@
-import 'package:dristi/src/core/styles/colors.dart';
+import 'package:dristi/src/core/styles/colors/colors.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
